+++ include/Debug_serial.h
@@ -14,12 +14,11 @@ namespace Debug_Serial {
     void println(const char* msg) { Serial.println(msg); }
     
     // Specialized throttle messages
-    void throttle_in(int pedal_filtered_1, int pedal_filtered_2, int pedal_filtered_final);
-    void throttle_out(float throttle_volt, int throttle_torque_val);
+    void throttle_in(uint16_t pedal_filtered_1, uint16_t pedal_filtered_2, uint16_t pedal_filtered_final);
+    void throttle_out(float throttle_volt, int16_t throttle_torque_val);
     void throttle_fault(Pedal_Fault_Status fault_status, float value);
     void status_car(CarStatus car_status);
-    void status_brake(int brake_voltage);
-    // #endif // DEBUG_SERIAL
+    void status_brake(float brake_voltage);
 }
 
 #endif // DEBUG_SERIAL_H
